delete test_tf2
