+++ tests/unit/examples/test_notebooks_python_tf2.py
@@ -16,24 +16,6 @@ ABS_TOL = 0.05
 
 
 @pytest.mark.notebooks
-def test_template_runs(notebooks, output_notebook, kernel_name):
-    notebook_path = notebooks["template"]
-    pm.execute_notebook(
-        notebook_path,
-        output_notebook,
-        parameters=dict(PM_VERSION=pm.__version__),
-        kernel_name=kernel_name,
-    )
-    nb = sb.read_notebook(output_notebook)
-    df = nb.papermill_dataframe
-    assert df.shape[0] == 2
-    check_version = df.loc[df["name"] == "checked_version", "value"].values[0]
-    assert check_version is True
-
-
-@pytest.mark.notebooks
 def test_sasrec_single_node_runs(notebooks, output_notebook, kernel_name):
     notebook_path = notebooks["sasrec_quickstart"]
     pm.execute_notebook(notebook_path, output_notebook, kernel_name=kernel_name)
-
-
